# More predictor optimizations