This gives another ~5% end-to-end improvement in lossless decoding.

Fixes #94 

 ## Microbenchmarks
Before:
```
test lossless_transform::benches::predictor10 ... bench:       1,295.97 ns/iter (+/- 27.69) = 787 MB/s
test lossless_transform::benches::predictor11 ... bench:       1,500.55 ns/iter (+/- 22.55) = 680 MB/s
test lossless_transform::benches::predictor12 ... bench:       1,172.88 ns/iter (+/- 21.36) = 870 MB/s
test lossless_transform::benches::predictor13 ... bench:       1,665.91 ns/iter (+/- 26.10) = 612 MB/s
```
After:

```
test lossless_transform::benches::predictor10 ... bench:         667.62 ns/iter (+/- 26.49) = 1529 MB/s
test lossless_transform::benches::predictor11 ... bench:         363.30 ns/iter (+/- 0.41) = 2809 MB/s
test lossless_transform::benches::predictor12 ... bench:         500.97 ns/iter (+/- 15.03) = 2040 MB/s
test lossless_transform::benches::predictor13 ... bench:         912.48 ns/iter (+/- 20.03) = 1118 MB/s
```